Rename existing helpers.ts file to utils.ts
