+++ src/logging.ts
@@ -1,4 +1,4 @@
-import { assertNever } from "./helpers"
+import { assertNever } from "./utils"
 
 export enum LogLevel {
   DEBUG = "debug",
